🔧 refactor: renamed folders (#686)
