+++ .vscode/settings.json
@@ -26,6 +26,7 @@
     "Multihash",
     "NSID",
     "Pageable",
+    "paginations",
     "pchar",
     "proto",
     "pubspec",
