+++ packages/atproto_core/lib/src/clients/retry_event.dart
