+++ packages/atproto_core/test/src/converters/at_uri_converter_test.dart
@@ -7,7 +7,7 @@ import 'package:at_uri/at_uri.dart';
 import 'package:test/test.dart';
 
 // 🌎 Project imports:
-import 'package:atproto_core/src/converter/at_uri_converter.dart';
+import 'package:atproto_core/src/converters/at_uri_converter.dart';
 
 const _value =
     'at://did:plc:jwq7l3xokmriwdz4fvqizvj5/app.bsky.feed.post/3k27ubjufzz2o';
